+++ src/main/java/net/minimaxi/voidminers/datagen/ModDataGen.java
@@ -12,7 +12,7 @@ import net.neoforged.neoforge.data.event.GatherDataEvent;
 import java.util.concurrent.CompletableFuture;
 
 @EventBusSubscriber(modid = VoidMiners.MODID)
-public class DataGenerators {
+public class ModDataGen {
     @SubscribeEvent
     public static void gatherData(GatherDataEvent event) {
         DataGenerator generator = event.getGenerator();