+++ src/main/java/net/minimaxi/voidminers/datagen/ModLootTableProvider.java
@@ -1,6 +1,5 @@
 package net.minimaxi.voidminers.datagen;
 
-import net.minimaxi.voidminers.datagen.loot.ModBlockLootTables;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.data.PackOutput;
 import net.minecraft.data.loot.LootTableProvider;