+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 1.21.1-1.6.6
 
 - Optimization of Miners and Solars
-- Changed the package name, if you are adding custom upgrade items, please see teh [wiki](https://github.com/Minimaxi28/VoidMiners-Remastered-Reforked/wiki#add-custom-max-storage-upgrades)
+- Changed the package name, if you are adding custom upgrade items, please see the [wiki](https://github.com/Minimaxi28/VoidMiners-Remastered-Reforked/wiki#add-custom-max-storage-upgrades)
 
 ## 1.21.1-1.6.5